Added NGRX store boilerplate code
Added constant file for certain urls
Fixed autoprefixer warning for leaflet as a temporary fix
Also manually changed it for the tailwind forms but if we don't use it for now we can hide them. It is however fixed in the latest version so if that releases, we could use it again.